Add improved card template image
Provide a better version of the card template as an image asset to replace or add to the project. The change supplies a new image file (img/2025-10-05_122459.png) representing the updated card template to support the PixiJS-to-React refactor described in the summary.

This commit addresses the request to create a better card template and supplies the new visual resource so the UI/refactor can use the improved design.